+++ tests/SoCreate.ServiceFabric.PubSub.Tests/GivenServiceReference.cs
@@ -54,5 +54,86 @@ namespace SoCreate.ServiceFabric.PubSub.Tests
             bool shouldDeliver = actorRef.ShouldDeliverMessage(messageWrapper);
             Assert.IsFalse(shouldDeliver);
         }
+
+        [TestMethod]
+        public void WhenDeterminingShouldDeliverMessageToActorWithMatchingPayload_ThenReturnsTrue()
+        {
+            var actorRef = new ActorReferenceWrapper(new ActorReference { ActorId = ActorId.CreateRandom() }, "Customer.Name=Customer1");
+            var messageWrapper = new
+            {
+                Customer = new
+                {
+                    Name = "Customer1"
+                }
+            }.CreateMessageWrapper();
+
+            bool shouldDeliver = actorRef.ShouldDeliverMessage(messageWrapper);
+            Assert.IsTrue(shouldDeliver);
+        }
+
+        [TestMethod]
+        public void WhenDeterminingShouldDeliverMessageToServiceWithMatchingPayloadWithRegex_ThenReturnsTrue()
+        {
+            var serviceRef = new ServiceReferenceWrapper(new ServiceReference(), "Customer.Name=^Customer");
+            var messageWrapper = new
+            {
+                Customer = new
+                {
+                    Name = "Customer1"
+                }
+            }.CreateMessageWrapper();
+
+            bool shouldDeliver = serviceRef.ShouldDeliverMessage(messageWrapper);
+            Assert.IsTrue(shouldDeliver);
+        }
+
+        [TestMethod]
+        public void WhenDeterminingShouldDeliverMessageToActorWithMatchingPayloadWithRegexReservedChar_ThenReturnsTrue()
+        {
+            var actorRef = new ActorReferenceWrapper(new ActorReference { ActorId = ActorId.CreateRandom() }, "Customer.Name=^Customer");
+            var messageWrapper = new
+            {
+                Customer = new
+                {
+                    Name = "Customer1"
+                }
+            }.CreateMessageWrapper();
+
+            bool shouldDeliver = actorRef.ShouldDeliverMessage(messageWrapper);
+            Assert.IsTrue(shouldDeliver);
+        }
+
+        [TestMethod]
+        public void WhenDeterminingShouldDeliverMessageToServiceWithUnMatchingPayloadWithRegexReservedChar_ThenReturnsFalse()
+        {
+            var serviceRef = new ServiceReferenceWrapper(new ServiceReference(), "Customer.Name=2$");
+            var messageWrapper = new
+            {
+                Customer = new
+                {
+                    Name = "Customer1"
+                }
+            }.CreateMessageWrapper();
+
+            bool shouldDeliver = serviceRef.ShouldDeliverMessage(messageWrapper);
+            Assert.IsFalse(shouldDeliver);
+        }
+
+        [TestMethod]
+        public void WhenDeterminingShouldDeliverMessageToActorWithUnMatchingPayloadWithRegexReservedChar_ThenReturnsFalse()
+        {
+            var actorRef = new ActorReferenceWrapper(new ActorReference { ActorId = ActorId.CreateRandom() }, "Customer.Name=2$");
+            var messageWrapper = new
+            {
+                Customer = new
+                {
+                    Name = "Customer1"
+                }
+            }.CreateMessageWrapper();
+
+            bool shouldDeliver = actorRef.ShouldDeliverMessage(messageWrapper);
+            Assert.IsFalse(shouldDeliver);
+        }
+
     }
 }